+++ libs/common/src/lib/interfaces/responses/public-portfolio-response.interface.ts
@@ -1,4 +1,5 @@
-import { PortfolioPosition } from '../portfolio-position.interface';
+import { PortfolioDetails, PortfolioPosition } from '..';
+import { Market } from '../../types';
 
 export interface PublicPortfolioResponse extends PublicPortfolioResponseV1 {
   alias?: string;
@@ -20,6 +21,12 @@ export interface PublicPortfolioResponse extends PublicPortfolioResponseV1 {
       | 'url'
       | 'valueInBaseCurrency'
       | 'valueInPercentage'
+    >;
+  };
+  markets: {
+    [key in Market]: Pick<
+      PortfolioDetails['markets'][key],
+      'id' | 'valueInPercentage'
     >;
   };
 }